[RESTEASY-1555] Move Serializable provider into org.jboss.resteasy.plugins.providers.legacy package
