+++ testsuite/integration-tests/src/test/resources/org/jboss/resteasy/test/resource/param/javax.ws.rs.ext.Providers
@@ -1,1 +1,1 @@
-org.jboss.resteasy.plugins.providers.SerializableProvider
+org.jboss.resteasy.plugins.providers.legacy.SerializableProvider
