+++ testsuite/legacy-integration-tests/src/test/java/org/jboss/resteasy/test/resource/param/resource/SerializableWithParametersResource.java
@@ -9,6 +9,6 @@ public class SerializableWithParametersResource {
     @GET
     @SuppressWarnings("deprecation")
     public Response test() {
-        return Response.ok().entity(new SerializableWithParametersObject("abc")).type(org.jboss.resteasy.plugins.providers.SerializableProvider.APPLICATION_SERIALIZABLE + ";q=0.5").build();
+        return Response.ok().entity(new SerializableWithParametersObject("abc")).type(org.jboss.resteasy.plugins.providers.legacy.SerializableProvider.APPLICATION_SERIALIZABLE + ";q=0.5").build();
     }
 }
